# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "API for easy study new things from school",
        "version": "0.3.4",
        "title": "Learn and test knowledge - API"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/quiz/administration/users": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin takes list of users which has account to delete",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "400": {
                        "description": "Error about data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/administration/schoolSubject": {
            "post": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can create new school subject",
                "requestBody": {
                    "description": "School subject data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_school_subject": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/administration/subject": {
            "delete": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can dalete school subject",
                "requestBody": {
                    "description": "School subject data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_school_subject": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/administration/chapter": {
            "delete": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can delete chapter",
                "requestBody": {
                    "description": "Chapter data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_chapter": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/administration/user": {
            "delete": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can delete user",
                "requestBody": {
                    "description": "Delete user by his id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "uuid"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/quiz/administration/allUsers": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can take list of all users",
                "responses": {
                    "200": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/administration/permissions": {
            "put": {
                "tags": [
                    "Administration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Admin can give teacher permissions",
                "requestBody": {
                    "description": "Give teacher permissions for user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "uuid"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}